lab_pdf_to_dataset/requirements.txt
pdfplumber
pandas
openpyxl
pytesseract
pillow
flask
flask-cors
python-docx
pywin32
